+++ crmeb/app/adminapi/controller/v1/setting/SystemRole.php
@@ -2,7 +2,7 @@
 // +----------------------------------------------------------------------
 // | CRMEB [ CRMEB赋能开发者，助力企业发展 ]
 // +----------------------------------------------------------------------
-// | Copyright (c) 2016~2020 https://www.crmeb.com All rights reserved.
+// | Copyright (c) 2016~2022 https://www.crmeb.com All rights reserved.
 // +----------------------------------------------------------------------
 // | Licensed CRMEB并不是自由软件，未经许可不能去掉CRMEB相关版权
 // +----------------------------------------------------------------------
@@ -11,6 +11,7 @@
 namespace app\adminapi\controller\v1\setting;
 
 use app\adminapi\controller\AuthController;
+use app\services\system\admin\SystemAdminServices;
 use app\services\system\admin\SystemRoleServices;
 use app\services\system\SystemMenusServices;
 use think\facade\App;
@@ -34,8 +35,7 @@ class SystemRole extends AuthController
 
     /**
      * 显示资源列表
-     *
-     * @return \think\Response
+     * @return mixed
      */
     public function index()
     {
@@ -48,9 +48,12 @@ class SystemRole extends AuthController
     }
 
     /**
-     * 显示创建资源表单页.
-     *
-     * @return \think\Response
+     * 显示创建资源表单页
+     * @param SystemMenusServices $services
+     * @return mixed
+     * @throws \think\db\exception\DataNotFoundException
+     * @throws \think\db\exception\DbException
+     * @throws \think\db\exception\ModelNotFoundException
      */
     public function create(SystemMenusServices $services)
     {
@@ -70,33 +73,36 @@ class SystemRole extends AuthController
             ['status', 0],
             ['checked_menus', [], '', 'rules']
         ]);
-        if (!$data['role_name']) return app('json')->fail('请输入身份名称');
+        if (!$data['role_name']) return app('json')->fail(400220);
         if (!is_array($data['rules']) || !count($data['rules']))
-            return app('json')->fail('请选择最少一个权限');
+            return app('json')->fail(400221);
         $data['rules'] = implode(',', $data['rules']);
         if ($id) {
-            if (!$this->services->update($id, $data)) return app('json')->fail('修改失败!');
+            if (!$this->services->update($id, $data)) return app('json')->fail(100007);
             \think\facade\Cache::clear();
-            return app('json')->success('修改成功!');
+            return app('json')->success(100001);
         } else {
             $data['level'] = $this->adminInfo['level'] + 1;
-            if (!$this->services->save($data)) return app('json')->fail('添加身份失败!');
+            if (!$this->services->save($data)) return app('json')->fail(400223);
             \think\facade\Cache::clear();
-            return app('json')->success('添加身份成功!');
+            return app('json')->success(400222);
         }
     }
 
     /**
-     * 显示编辑资源表单页.
-     *
-     * @param int $id
-     * @return \think\Response
+     * 显示编辑资源表单页
+     * @param SystemMenusServices $services
+     * @param $id
+     * @return mixed
+     * @throws \think\db\exception\DataNotFoundException
+     * @throws \think\db\exception\DbException
+     * @throws \think\db\exception\ModelNotFoundException
      */
     public function edit(SystemMenusServices $services, $id)
     {
         $role = $this->services->get($id);
         if (!$role) {
-            return app('json')->fail('修改的规格不存在');
+            return app('json')->fail(100100);
         }
         $menus = $services->getMenus($this->adminInfo['level'] == 0 ? [] : $this->adminInfo['roles']);
         return app('json')->success(['role' => $role->toArray(), 'menus' => $menus]);
@@ -104,17 +110,20 @@ class SystemRole extends AuthController
 
     /**
      * 删除指定资源
-     *
-     * @param int $id
-     * @return \think\Response
+     * @param SystemAdminServices $adminServices
+     * @param $id
+     * @return mixed
      */
-    public function delete($id)
+    public function delete(SystemAdminServices $adminServices, $id)
     {
+        if ($adminServices->checkRoleUse($id)) {
+            return app('json')->fail(400754);
+        }
         if (!$this->services->delete($id))
-            return app('json')->fail('删除失败,请稍候再试!');
+            return app('json')->fail(100008);
         else {
             \think\facade\Cache::clear();
-            return app('json')->success('删除成功!');
+            return app('json')->success(100002);
         }
     }
 
@@ -127,18 +136,18 @@ class SystemRole extends AuthController
     public function set_status($id, $status)
     {
         if (!$id) {
-            return app('json')->fail('缺少参数');
+            return app('json')->fail(100100);
         }
         $role = $this->services->get($id);
         if (!$role) {
-            return app('json')->fail('没有查到此身份');
+            return app('json')->fail(400199);
         }
         $role->status = $status;
         if ($role->save()) {
             \think\facade\Cache::clear();
-            return app('json')->success('修改成功');
+            return app('json')->success(100001);
         } else {
-            return app('json')->fail('修改失败');
+            return app('json')->fail(100007);
         }
     }
 }
